+++ media/js/admin/file.js
@@ -9,25 +9,6 @@ jQuery(document).ready(function() {
 
     jQuery("#js-iproof-btn-download").on("click", function(event){
         event.preventDefault();
-
-        var fields = {
-            task: "file.getFormToken",
-            format: "raw"
-        };
-
-        jQuery.ajax({
-            url: "index.php?option=com_identityproof",
-            type: "get",
-            data: fields,
-            dataType: "text json"
-        }).done(function (response) {
-
-            if (response.success) {
-                jQuery("#js-iproof-download-token").attr("name", response.data.token);
-                jQuery("#js-iproof-download-form").submit();
-            }
-
-        });
-
+        jQuery("#js-iproof-download-form").submit();
     });
 });