+++ plugins/content/plg_content_identityproofusermail/identityproofusermail.php
@@ -38,9 +38,8 @@ class plgContentIdentityProofUserMail extends JPlugin
     }
 
     /**
-     * Send notification mail to a user when his project be approved.
-     * If I return NULL, an message will not be displayed in the browser.
-     * If I return FALSE, an error message will be displayed in the browser.
+     * Send notification mail to a user
+     * when the administrator set his state to VERIFIED.
      *
      * @param string $context
      * @param array $ids
@@ -64,12 +63,11 @@ class plgContentIdentityProofUserMail extends JPlugin
         // Initialize plugin
         $this->init();
 
-        // Check for enabled option for sending mail
-        // when administrator approve project.
+        // Email template ID.
         $emailId = $this->params->get("send_when_verified", 0);
         if (!$emailId) {
-            JLog::add(JText::sprintf("PLG_CONTENT_IDENTITYPROOFUSERMAIL_ERROR_INVALID_EMAIL_TEMPLATE", $this->name));
-            return false;
+            JLog::add(JText::sprintf("PLG_CONTENT_IDENTITYPROOFUSERMAIL_ERROR_INVALID_EMAIL_TEMPLATE", $this->name), JLog::DEBUG);
+            return null;
         }
 
         Joomla\Utilities\ArrayHelper::toInteger($ids);
@@ -80,14 +78,14 @@ class plgContentIdentityProofUserMail extends JPlugin
             $users->load(array("ids" => $ids));
 
             if (0 >= count($users)) {
-                JLog::add(JText::sprintf("PLG_CONTENT_IDENTITYPROOFUSERMAIL_ERROR_INVALID_USERS_S", $this->name));
-                return false;
+                JLog::add(JText::sprintf("PLG_CONTENT_IDENTITYPROOFUSERMAIL_ERROR_INVALID_USERS_S", $this->name), JLog::ERROR);
+                return null;
             }
 
             foreach ($users as $user) {
 
                 // Send email to users.
-                $return = $this->sendMail($user, $emailId);
+                $return = $this->sendMail($user, $emailId, "user");
 
                 // If there is an error, stop the loop.
                 // Let the administrator to look the errors.
@@ -102,7 +100,71 @@ class plgContentIdentityProofUserMail extends JPlugin
         return true;
     }
 
-    protected function sendMail($user, $emailId)
+    /**
+     * Send notification mail to a user
+     * when the administrator leave a notice about a file.
+     *
+     * @param string $context
+     * @param int $id File ID
+     *
+     * @return bool|null
+     */
+    public function onContentLeaveNote($context, $id)
+    {
+        $app = JFactory::getApplication();
+        /** @var $app JApplicationSite */
+
+        if (!$app->isAdmin()) {
+            return null;
+        }
+
+        if (strcmp("com_identityproof.notification", $context) != 0) {
+            return null;
+        }
+
+        // Initialize plugin
+        $this->init();
+
+        // Email template ID.
+        $emailId = $this->params->get("send_when_leave_notice", 0);
+        if (!$emailId) {
+            JLog::add(JText::sprintf("PLG_CONTENT_IDENTITYPROOFUSERMAIL_ERROR_INVALID_EMAIL_TEMPLATE", $this->name), JLog::DEBUG);
+            return null;
+        }
+
+        if (!empty($id)) {
+
+            $file = new IdentityProof\File(JFactory::getDbo());
+            $file->load($id);
+
+            if (!$file->getId()) {
+                JLog::add(JText::sprintf("PLG_CONTENT_IDENTITYPROOFUSERMAIL_ERROR_INVALID_FILE_S", $this->name), JLog::ERROR);
+                return null;
+            }
+
+            // Send email to users.
+            $return = $this->sendMail($file, $emailId, "file");
+
+            // If there is an error, stop the loop.
+            // Let the administrator to look the errors.
+            if ($return !== true) {
+                return false;
+            }
+
+        }
+
+        return true;
+    }
+
+    /**
+     * @param object $item
+     * @param int $emailId
+     * @param string $type
+     *
+     * @return bool
+     * @throws Exception
+     */
+    protected function sendMail($item, $emailId, $type)
     {
         $app = JFactory::getApplication();
         /** @var $app JApplicationSite */
@@ -143,11 +205,29 @@ class plgContentIdentityProofUserMail extends JPlugin
             }
         }
 
-        // Prepare data for parsing
         $data["administrator_name"]     = $email->getSenderName();
         $data["administrator_email"]    = $email->getSenderEmail();
-        $data["user_name"]              = $user["name"];
-        $data["user_email"]             = $user["email"];
+
+        // Prepare data for parsing.
+        switch ($type) {
+
+            case "file":
+
+                $user = JFactory::getUser($item->getUserId());
+
+                $data["user_name"]      = $user->get("name");
+                $data["user_email"]     = $user->get("email");
+                $data["note"]           = $item->getNote();
+                $data["filename"]       = $item->getFilename();
+
+                break;
+
+            case "user":
+                $data["user_name"]      = $item["name"];
+                $data["user_email"]     = $item["email"];
+                break;
+
+        }
 
         $email->parse($data);
         $subject = $email->getSubject();
@@ -155,9 +235,9 @@ class plgContentIdentityProofUserMail extends JPlugin
 
         $mailer = JFactory::getMailer();
         if (strcmp("html", $emailMode) == 0) { // Send as HTML message
-            $result = $mailer->sendMail($email->getSenderEmail(), $email->getSenderName(), $data["file_owner_email"], $subject, $body, Prism\Constants::MAIL_MODE_HTML);
+            $result = $mailer->sendMail($email->getSenderEmail(), $email->getSenderName(), $data["user_email"], $subject, $body, Prism\Constants::MAIL_MODE_HTML);
         } else { // Send as plain text.
-            $result = $mailer->sendMail($email->getSenderEmail(), $email->getSenderName(), $data["file_owner_email"], $subject, $body, Prism\Constants::MAIL_MODE_PLAIN);
+            $result = $mailer->sendMail($email->getSenderEmail(), $email->getSenderName(), $data["user_email"], $subject, $body, Prism\Constants::MAIL_MODE_PLAIN);
         }
 
         // Log the error.