+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Proof Of Identity Release Notes
 ===============================
 
+###v2.1
+* Fixed the plugin "Content - Proof of Identity User Mail".
+* Added option to send mails when the administrator leave a note to a file.
+
 ###v2.0
 * Done to be based on Bootstrap 3.x
 * Refactored the library.